Move SearchListAdapter to adapter package
