+++ app/src/main/java/de/ur/explure/adapter/SearchListAdapter.kt
@@ -1,4 +1,4 @@
-package de.ur.explure
+package de.ur.explure.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -9,7 +9,7 @@ import de.ur.explure.databinding.SearchItemBinding
 import de.ur.explure.model.route.Route
 
 class SearchListAdapter(private val onClick: (Route) -> Unit) :
-    ListAdapter<Route, SearchListAdapter.RouteViewHolder>(RouteDiffCallback) {
+    ListAdapter<Route, SearchListAdapter.RouteViewHolder>(SearchListDiffCallback) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RouteViewHolder {
         return RouteViewHolder.from(parent)
@@ -40,7 +40,7 @@ class SearchListAdapter(private val onClick: (Route) -> Unit) :
     }
 }
 
-object RouteDiffCallback : DiffUtil.ItemCallback<Route>() {
+object SearchListDiffCallback : DiffUtil.ItemCallback<Route>() {
     override fun areItemsTheSame(oldItem: Route, newItem: Route): Boolean {
         return oldItem == newItem
     }
